+++ tests/test_configparser.py
@@ -1,30 +1,30 @@
 from pathlib import Path
 
 from clickgen.configparser import (
-    parse_toml_config_section,
+    parse_config_section,
+    parse_theme_section,
     parse_toml_file,
-    parse_toml_theme_section,
 )
 
 td = {"theme": {"name": "test", "comment": "test", "website": "test"}}
 
 
-def test_parse_toml_theme_section():
-    t = parse_toml_theme_section(td)
+def test_parse_theme_section():
+    t = parse_theme_section(td)
     assert t.name == "test"
     assert t.comment == "test"
     assert t.website == "test"
 
 
-def test_parse_toml_theme_section_with_kwargs():
+def test_parse_theme_section_with_kwargs():
     kwargs = {"name": "new", "comment": "new", "website": "new"}
-    t = parse_toml_theme_section(td, **kwargs)
+    t = parse_theme_section(td, **kwargs)
     assert t.name == kwargs["name"]
     assert t.comment == kwargs["comment"]
     assert t.website == kwargs["website"]
 
 
-dd = {
+dd1 = {
     "config": {
         "bitmaps_dir": "test",
         "out_dir": "test",
@@ -35,8 +35,37 @@ dd = {
 }
 
 
-def test_parse_toml_config_section():
-    c = parse_toml_config_section("", dd)
+def test_win_size_deprecation_message(capsys):
+    parse_config_section("", dd1)
+
+    captured = capsys.readouterr()
+    assert (
+        "Warning: Option 'win_size' is deprecated. Use 'win_sizes' inside individual cursor settings or set to 'cursor.fallback'"
+        in captured.out
+    )
+
+
+def test_x11_sizes_deprecation_message(capsys):
+    parse_config_section("", dd1)
+
+    captured = capsys.readouterr()
+    assert (
+        "Warning: Option 'x11_size' is deprecated. Use 'x11_sizes' inside individual cursor settings or set to 'cursor.fallback'"
+        in captured.out
+    )
+
+
+dd2 = {
+    "config": {
+        "bitmaps_dir": "test",
+        "out_dir": "test",
+        "platforms": "test",
+    }
+}
+
+
+def test_parse_config_section():
+    c = parse_config_section("", dd2)
     assert isinstance(c.bitmaps_dir, Path)
     assert c.bitmaps_dir.name is "test"
     assert c.bitmaps_dir.is_absolute()
@@ -45,42 +74,35 @@ def test_parse_toml_config_section():
     assert c.out_dir.is_absolute()
 
     assert c.platforms == "test"
-    assert c.x11_sizes == 10
-    assert c.win_size == 11
 
 
-def test_parse_toml_config_section_with_absolute_paths():
-    dd["config"]["bitmaps_dir"] = Path("test").absolute()
-    dd["config"]["out_dir"] = Path("test").absolute()
+def test_parse_config_section_with_absolute_paths():
+    dd2["config"]["bitmaps_dir"] = str(Path("test").absolute())
+    dd2["config"]["out_dir"] = str(Path("test").absolute())
 
-    c = parse_toml_config_section("", dd)
+    c = parse_config_section("", dd2)
 
     assert isinstance(c.bitmaps_dir, Path)
     assert c.bitmaps_dir.is_absolute()
-    assert str(c.bitmaps_dir) == str(dd["config"]["bitmaps_dir"])
+    assert str(c.bitmaps_dir) == dd2["config"]["bitmaps_dir"]
     assert isinstance(c.out_dir, Path)
-    assert str(c.out_dir) == str(dd["config"]["out_dir"])
+    assert str(c.out_dir) == dd2["config"]["out_dir"]
 
 
-def test_parse_toml_config_section_with_kwargs():
-
+def test_parse_config_section_with_kwargs():
     kwargs = {
         "bitmaps_dir": "new",
         "out_dir": "new",
         "platforms": "new",
-        "x11_sizes": [100, 10],
-        "win_size": 110,
     }
 
-    c = parse_toml_config_section("", dd, **kwargs)
+    c = parse_config_section("", dd2, **kwargs)
     assert c.bitmaps_dir == kwargs["bitmaps_dir"]
     assert c.out_dir == kwargs["out_dir"]
     assert c.platforms == kwargs["platforms"]
-    assert c.x11_sizes == kwargs["x11_sizes"]
-    assert c.win_size == kwargs["win_size"]
 
 
-def test_parse_toml_file(samples_dir: Path):
+def test_parse_file(samples_dir: Path):
     fp = samples_dir / "sample.toml"
     c = parse_toml_file(str(fp.absolute()))
     assert c.cursors[0].win_cursor_name == "Default.cur"